+++ partner/src/epilot/models/operations/searchassignable.py
@@ -3,10 +3,13 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
+from ..shared import assignableorganization as shared_assignableorganization
+from ..shared import assignablepartneruser as shared_assignablepartneruser
+from ..shared import assignableuser as shared_assignableuser
 from dataclasses_json import Undefined, dataclass_json
 from enum import Enum
 from epilot import utils
-from typing import Any, Optional
+from typing import Optional, Union
 
 class SearchAssignableRequestBodyTypes(str, Enum):
     USER = 'user'
@@ -39,7 +42,7 @@ class SearchAssignable200ApplicationJSON:
     r"""List of assignable results"""
     hits: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('hits'), 'exclude': lambda f: f is None }})
     r"""total number of search results"""
-    results: Optional[list[Any]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('results'), 'exclude': lambda f: f is None }})
+    results: Optional[list[Union[shared_assignableuser.AssignableUser, shared_assignablepartneruser.AssignablePartnerUser, shared_assignableorganization.AssignableOrganization]]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('results'), 'exclude': lambda f: f is None }})